trading-agent/requirements.txt
requests>=2.28.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
anthropic>=0.20.0
pybit>=5.10.0
